feature: configure table column visibility (#623)
