+++ src/components/table/use-cases/traces.tsx
@@ -4,6 +4,7 @@ import { DataTable } from "@/src/components/table/data-table";
 import { DataTableToolbar } from "@/src/components/table/data-table-toolbar";
 import TableLink from "@/src/components/table/table-link";
 import { TokenUsageBadge } from "@/src/components/token-usage-badge";
+import useColumnVisibility from "@/src/features/column-visibility/hooks/useColumnVisibility";
 import { useQueryFilterState } from "@/src/features/filters/hooks/useFilterState";
 import { type FilterState } from "@/src/features/filters/types";
 import { useDetailPageLists } from "@/src/features/navigate-detail-pages/context";
@@ -23,7 +24,7 @@ import {
 } from "use-query-params";
 import { BookmarkTrace } from "@/src/components/bookmark-trace";
 
-export type TraceTableRow = {
+export type TracesTableRow = {
   bookmarked: boolean;
   id: string;
   timestamp: string;
@@ -41,7 +42,7 @@ export type TraceTableRow = {
   };
 };
 
-export type TraceTableProps = {
+export type TracesTableProps = {
   projectId: string;
   userId?: string;
   omittedFilter?: string[];
@@ -53,13 +54,12 @@ export default function TracesTable({
   projectId,
   userId,
   omittedFilter = [],
-}: TraceTableProps) {
+}: TracesTableProps) {
   const { setDetailPageList } = useDetailPageLists();
   const [searchQuery, setSearchQuery] = useQueryParam(
     "search",
     withDefault(StringParam, null),
   );
-
   const [userFilterState, setUserFilterState] = useQueryFilterState([
     {
       column: "timestamp",
@@ -81,7 +81,6 @@ export default function TracesTable({
     : [];
 
   const filterState = userFilterState.concat(userIdFilter);
-
   const [paginationState, setPaginationState] = useQueryParams({
     pageIndex: withDefault(NumberParam, 0),
     pageSize: withDefault(NumberParam, 50),
@@ -111,7 +110,7 @@ export default function TracesTable({
 
   const convertToTableRow = (
     trace: RouterOutput["traces"]["all"][0],
-  ): TraceTableRow => {
+  ): TracesTableRow => {
     return {
       bookmarked: trace.bookmarked,
       id: trace.id,
@@ -131,7 +130,7 @@ export default function TracesTable({
     };
   };
 
-  const columns: ColumnDef<TraceTableRow>[] = [
+  const columns: ColumnDef<TracesTableRow>[] = [
     {
       accessorKey: "bookmarked",
       header: undefined,
@@ -162,14 +161,17 @@ export default function TracesTable({
           />
         ) : undefined;
       },
+      enableHiding: true,
     },
     {
       accessorKey: "timestamp",
       header: "Timestamp",
+      enableHiding: true,
     },
     {
       accessorKey: "name",
       header: "Name",
+      enableHiding: true,
     },
     {
       accessorKey: "userId",
@@ -185,6 +187,7 @@ export default function TracesTable({
           />
         ) : undefined;
       },
+      enableHiding: true,
     },
     {
       accessorKey: "latency",
@@ -194,6 +197,7 @@ export default function TracesTable({
         const value: number | undefined = row.getValue("latency");
         return value !== undefined ? `${value.toFixed(2)} sec` : undefined;
       },
+      enableHiding: true,
     },
     {
       accessorKey: "usage",
@@ -213,6 +217,7 @@ export default function TracesTable({
           />
         );
       },
+      enableHiding: true,
     },
     {
       accessorKey: "scores",
@@ -222,6 +227,7 @@ export default function TracesTable({
         const values: Score[] = row.getValue("scores");
         return <GroupedScoreBadges scores={values} variant="headings" />;
       },
+      enableHiding: true,
     },
     {
       accessorKey: "metadata",
@@ -230,14 +236,17 @@ export default function TracesTable({
         const values: string = row.getValue("metadata");
         return <div className="flex flex-wrap gap-x-3 gap-y-1">{values}</div>;
       },
+      enableHiding: true,
     },
     {
       accessorKey: "version",
       header: "Version",
+      enableHiding: true,
     },
     {
       accessorKey: "release",
       header: "Release",
+      enableHiding: true,
     },
     {
       accessorKey: "action",
@@ -252,12 +261,17 @@ export default function TracesTable({
           />
         ) : undefined;
       },
+      enableHiding: true,
     },
   ];
+
+  const [columnVisibility, setColumnVisibility] =
+    useColumnVisibility<TracesTableRow>("tracesColumnVisibility", columns);
 
   return (
     <div>
       <DataTableToolbar
+        columns={columns}
         filterColumnDefinition={tracesTableColsWithOptions(
           traceFilterOptions.data,
         )}
@@ -268,6 +282,8 @@ export default function TracesTable({
         }}
         filterState={userFilterState}
         setFilterState={setUserFilterState}
+        columnVisibility={columnVisibility}
+        setColumnVisibility={setColumnVisibility}
       />
       <DataTable
         columns={columns}
@@ -291,6 +307,8 @@ export default function TracesTable({
           onChange: setPaginationState,
           state: paginationState,
         }}
+        columnVisibility={columnVisibility}
+        onColumnVisibilityChange={setColumnVisibility}
       />
     </div>
   );
